import now requires strings import checks builtin values before external files

+++ src/errors.d
@@ -30,7 +30,7 @@ class OratrInvalidCodeException : OratrBaseException {
 }
 
 class OratrArgumentCountException : OratrBaseException {
-	this( uint count, string func, string possible, string file = __FILE__, size_t line = __LINE__ )
+	this( size_t count, string func, string possible, string file = __FILE__, size_t line = __LINE__ )
     {
         super( "Invalid argument count "~to!string(count)~" in function "~func
         	~" accepting "~to!string(possible)~" arguments",
@@ -39,7 +39,7 @@ class OratrArgumentCountException : OratrBaseException {
 }
 
 class OratrArgumentCountExceptionControlStructure : OratrBaseException {
-	this( uint count, string func, string possible, string file = __FILE__, size_t line = __LINE__ )
+	this( size_t count, string func, string possible, string file = __FILE__, size_t line = __LINE__ )
     {
         super( "Invalid argument count "~to!string(count)~" in control structure "~func
         	~" accepting "~to!string(possible)~" arguments",
